refactor(analytics): split mlTelemetry/trackers into thematic modules
Decompose the 1008-LOC trackers.ts into 6 sibling files organized by the
event family they emit. trackers.ts becomes a 44-LOC re-export barrel so
the public surface (init.ts, index.ts, init.test.ts via wildcard import)
is byte-identical.

Extracted modules (all under 300 LOC):

- trackersHelpers.ts (46) — shared `isEnabled` gate + the adjacency
  helper consumed by placement tracking.
- trackersBinPlacement.ts (142) — `trackBinPlacement` (with
  high-volume sampling for unlabeled bins) and `trackBulkPlacement`.
- trackersBinTransform.ts (201) — `trackBinResize`, `trackBinDeletion`
  (with the bin_abandoned special case for short-lived unlabeled bins),
  `trackBinMove`, `trackBinRotation`.
- trackersBinMetadata.ts (86) — `trackLabelUpdate`, `trackCategoryChange`.
- trackersLayout.ts (260) — `trackLayoutSnapshot` (rate-limited),
  `trackDrawerPurpose`, `trackDrawerResize`, `trackFillOperation`,
  `trackLayerMove`, plus the internal `trackCrossLayoutPattern` emitter
  scheduled from snapshot commits at permanent triggers (save / share /
  session_end). Internal so the schedule from snapshot remains the only
  call site.
- trackersQuality.ts (281) — `trackQualitySignal`,
  `trackPlacementRejection`, `trackUndo`, `trackQuickCorrection`,
  `trackSessionSummary`.

trackers.ts: 1008 LOC -> 44 LOC. All 3 test files / 181 tests pass
(784ms targeted run).
